refactor: move all studentForm hooks to useStudentForm custom hook
